aws-lambda/imgmod/requirements.txt
boto3
requests
pillow
